move BigTensor out of kernels
closes #15
